feat: enhance UI components and implement NotFound page, improve authentication handling and routing

+++ BlogApp/src/components/PostForm.jsx
@@ -17,25 +17,23 @@ const PostForm = ({ post }) => {
     });
 
   const navigate = useNavigate();
-  const userData = useSelector((state) => state.user.userData);
+  const userData = useSelector((state) => state.auth.userData);
   const submit = async (data) => {
     if (post) {
-      const file = data.image[0]
-        ? appwriteService.uploadFile(DisposableStack.image)
-        : null;
-
-      if (file) {
-        appwriteService.deleteFile(post.image);
+      let file = null;
+      if (data.image && data.image[0]) {
+        file = await appwriteService.uploadFile(data.image[0]);
+        if (file) {
+          await appwriteService.deleteFile(post.image);
+        }
       }
-
       const dbPost = await appwriteService.updatePost(post.$id, {
         ...data,
-        image: file ? file.$id : undefined,
-
-        if(dbPost) {
-          navigate(`/post/${post.$id}`);
-        },
+        image: file ? file.$id : post.image,
       });
+      if (dbPost) {
+        navigate(`/post/${post.$id}`);
+      }
     } else {
       const file = await appwriteService.uploadFile(data.image[0]);
       if (file) {
@@ -45,7 +43,6 @@ const PostForm = ({ post }) => {
           ...data,
           userId: userData.$id,
         });
-
         if (dbPost) {
           navigate(`/post/${dbPost.$id}`);
         }
